backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
timm
transformers
pillow
numpy
scikit-learn
matplotlib
opencv-python
scikit-image
lime
